Rename genio brightness so that it doesn't clash
